pyproject.toml
[project]
name = "fauxdata"
version = "0.1.0"
description = "CLI for generating and validating fake datasets"
requires-python = ">=3.11"
dependencies = [
    "pointblank>=0.22",
    "typer>=0.12",
    "rich>=13",
    "pyfiglet>=1.0",
    "questionary>=2.0",
    "polars>=1.0",
    "pyyaml>=6.0",
    "faker>=26.0",
]

[project.scripts]
fauxdata = "fauxdata.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fauxdata"]
